Renamed util to common as it conflicts with the util on the toplevel directory.
