+++ web/vehicles/models.py
@@ -12,7 +12,7 @@ from django.db import models
 from django.core.exceptions import ValidationError
 from djgeojson.fields import PointField
 from django.conf import settings
-from util.user import Account
+from common.user import Account
 from security.models import JSONWebKey
 
 class RVIStatus:
